+++ src/css/styles.css
@@ -28,10 +28,10 @@ and https://www.w3schools.com/howto/howto_js_slideshow.asp
   --error: #fb36408e;
 
   /* mobile screen font sizes */
-  --font-h1: 3rem;
-  --font-h2: 2rem;
+  --font-h1: 3.5rem;
+  --font-h2: 2.5rem;
   --font-h3: 1.5rem;
-  --font-p: 1rem;
+  --font-p: 1em;
   --small-font: 0.8rem;
 
   --border-radius: 20px;
@@ -82,10 +82,12 @@ body {
   font-family: var(--font-body);
   font-size: var(--font-p);
   color: var(--text-color);
-  background: radial-gradient(circle at 50% 50%,
-      var(--main-grad1),
-      var(--main-grad2),
-      var(--main-grad3));
+  background: radial-gradient(
+    circle at 50% 50%,
+    var(--main-grad1),
+    var(--main-grad2),
+    var(--main-grad3)
+  );
 }
 
 #content {
@@ -123,10 +125,12 @@ body {
   text-decoration: none;
   padding: 8px;
   margin-bottom: 7px;
-  background: radial-gradient(circle at 50% 50%,
-      var(--nav-grad1),
-      var(--nav-grad2),
-      var(--main-grad3));
+  background: radial-gradient(
+    circle at 50% 50%,
+    var(--nav-grad1),
+    var(--nav-grad2),
+    var(--main-grad3)
+  );
   box-shadow:
     0 1px 3px var(--main-grad1),
     0 1px 2px var(--main-lighter);
@@ -153,10 +157,12 @@ body {
   text-decoration: none;
   padding: 8px;
   margin-bottom: 7px;
-  background: radial-gradient(circle at 50% 50%,
-      var(--nav-grad1),
-      var(--nav-grad2),
-      var(--main-grad3));
+  background: radial-gradient(
+    circle at 50% 50%,
+    var(--nav-grad1),
+    var(--nav-grad2),
+    var(--main-grad3)
+  );
   box-shadow:
     0 1px 3px var(--main-grad1),
     0 1px 2px var(--main-lighter);
@@ -197,7 +203,8 @@ body {
 }
 
 /* Styling for pokemon type cards on home page*/
-.main-right {
+.main-right,
+.pokemon-by-type {
   padding: 15px;
   color: var(--text-color);
   text-decoration: none;
@@ -212,7 +219,8 @@ body {
   text-align: center;
 }
 
-.main-right img {
+.main-right img,
+.pokemon-by-type img {
   box-shadow:
     0 2px 6px var(--accent-yellow),
     0 2px 4px var(--accent-blue);
@@ -221,7 +229,8 @@ body {
 }
 
 /* Peter is the best for figuring this out! */
-.main-right ul {
+.main-right ul,
+.pokemon-by-type ul {
   display: grid;
   grid-template-columns: 1fr 1fr;
   list-style-type: none;
@@ -262,9 +271,11 @@ https://www.geeksforgeeks.org/create-a-css-fireflies-background-using-html-css/
   /*glowing effect for the fireflies: use a yellow background color 
   along with radial-gradient and box shadow */
   background-color: rgb(255, 255, 73);
-  background-image: radial-gradient(rgb(249, 206, 36) 5%,
-      rgb(254, 179, 4) 25%,
-      rgb(252, 191, 7) 60%);
+  background-image: radial-gradient(
+    rgb(249, 206, 36) 5%,
+    rgb(254, 179, 4) 25%,
+    rgb(252, 191, 7) 60%
+  );
   box-shadow:
     0 0 5px 2px rgb(250, 193, 93),
     0 0 20px 10px rgb(255, 228, 140),
@@ -323,7 +334,6 @@ https://www.geeksforgeeks.org/create-a-css-fireflies-background-using-html-css/
 /* Use step by step by step ANIMATION to create the fireflies movement */
 /* create left/right movement */
 @keyframes leftright {
-
   0%,
   100% {
     left: 80%;
@@ -352,7 +362,6 @@ https://www.geeksforgeeks.org/create-a-css-fireflies-background-using-html-css/
 
 /* create up/down movement */
 @keyframes updown {
-
   0%,
   100% {
     top: 10px;
@@ -373,7 +382,6 @@ https://www.geeksforgeeks.org/create-a-css-fireflies-background-using-html-css/
 
 /* create blinking effect */
 @keyframes blinking {
-
   0%,
   100% {
     box-shadow:
@@ -408,10 +416,12 @@ https://www.geeksforgeeks.org/create-a-css-fireflies-background-using-html-css/
   text-align: right;
   color: var(--text-color);
   font-weight: bold;
-  background: radial-gradient(circle at 50% 50%,
-      var(--nav-grad1),
-      var(--nav-grad2),
-      var(--main-grad3));
+  background: radial-gradient(
+    circle at 50% 50%,
+    var(--nav-grad1),
+    var(--nav-grad2),
+    var(--main-grad3)
+  );
   padding: 2em;
 }
 
@@ -545,6 +555,16 @@ input:required:valid {
   text-transform: uppercase;
 }
 
+#register a {
+  color: var(--main-grad3);
+}
+
+.reg-now {
+  color: var(--text-color);
+  font-size: var(--font-p);
+  padding: 1em;
+}
+
 /* Button styling */
 .button {
   background-color: var(--accent-yellow);
@@ -563,15 +583,7 @@ input:required:valid {
   background-color: var(--accent-green);
 }
 
-#register a {
-  color: var(--main-grad3);
-}
-
-.reg-now{
-  color: var(--text-color);
-  font-size: var(--font-p);
-  padding: 1em;
-}
+/* Pokemon By Type pages to display each type*/
 
 /* 
 ******* ******* ******* ******* ******* ******* ******* *******
@@ -582,10 +594,10 @@ input:required:valid {
 */
 @media screen and (min-width: 600px) {
   :root {
-    --font-h1: 4rem;
-    --font-h2: 2.5rem;
-    --font-h3: 2rem;
-    --font-p: 1.2rem;
+    --font-h1: 4.5rem;
+    --font-h2: 3.5rem;
+    --font-h3: 2.5rem;
+    --font-p: 1.5rem;
   }
 
   /* https://www.w3schools.com/css/css_rwd_intro.asp */
@@ -645,7 +657,8 @@ input:required:valid {
     width: 15%;
   }
 
-  .side-column {}
+  .side-column {
+  }
 
   /* Peter is the best for figuring this out! */
   .main-right ul {
@@ -658,20 +671,20 @@ input:required:valid {
   }
 
   /*Used tutorial from W3SCHOOLS TO CREATE THE RESPONSIVE LAYOUT: https://www.w3schools.com/css/tryit.asp?filename=trycss_form_responsive */
-  input[type=text],
-  input[type=email],
-  input[type=password],
+  input[type="text"],
+  input[type="email"],
+  input[type="password"],
   select,
   textarea {
     width: 100%;
     padding: 1em;
     border: 1px solid var(--accent-green);
     resize: vertical;
-    color:black;
+    color: black;
   }
 
   label {
-    padding: .5em;
+    padding: 0.5em;
     display: inline-block;
   }
 
@@ -692,5 +705,4 @@ input:required:valid {
     display: table;
     clear: both;
   }
-
 }
