+++ src/js/main.js
@@ -1,7 +1,6 @@
 import pokemon from "pokemontcgsdk";
 import { loadHeaderFooter } from "./utils.mjs";
 
-
 const apiKey = import.meta.env.VITE_API_KEY;
 
 pokemon.configure({
