+++ src/login/index.html
@@ -1,98 +1,116 @@
 <!doctype html>
 <html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Yanez | Login</title>
+    <link rel="icon" type="image/png" href="/public/images/pokemonIcon.png" />
+    <link rel="stylesheet" href="/css/styles.css" />
+  </head>
 
-<head>
-  <meta charset="UTF-8" />
-  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Yanez | Login</title>
-  <link rel="stylesheet" href="/css/styles.css" />
-</head>
-
-<body>
-  <div id="content">
-    <div class="h1-p1">
-      <h1>Pocket Monsters</h1>
-      <p class="nav-p" id="nav-p"></p>
-    </div>
-    <div class="row">
-      <header class="divider" id="main-header"></header>
-      <div class="firefly" id="firefly">
-        <ul class="fireflies">
-          <li></li>
-          <li></li>
-          <li></li>
-          <li></li>
-          <li></li>
-          <li></li>
-          <li></li>
-        </ul>
+  <body>
+    <div id="content">
+      <div class="h1-p1">
+        <h1>Pocket Monsters</h1>
+        <p class="nav-p" id="nav-p"></p>
       </div>
-      <div class="main">
-        <div class="side col-6 col-s-9">
-          <h2>Hello Trainer!</h2>
-          <div class="login-side-column">
-            <p>
-              All is fair in love, war, and Pokémon battles.
-            </p>
-          </div>
-          <div class="main-column">
-            <div class="col-3 col-s-12">
-              <div class="main-right">
-                <div class="form-contain">
-                  <form class="loginForm" id="loginForm">
-                    <div class="form-row">
-                      <div class="form-col-25">
-                        <label for="userName">User Name:</label>
+      <div class="row">
+        <header class="divider" id="main-header"></header>
+        <div class="firefly" id="firefly">
+          <ul class="fireflies">
+            <li></li>
+            <li></li>
+            <li></li>
+            <li></li>
+            <li></li>
+            <li></li>
+            <li></li>
+          </ul>
+        </div>
+        <div class="main">
+          <div class="side col-6 col-s-9">
+            <h2>Hello Trainer!</h2>
+            <div class="login-side-column">
+              <p>All is fair in love, war, and Pokémon battles.</p>
+            </div>
+            <div class="main-column">
+              <div class="col-3 col-s-12">
+                <div class="main-right">
+                  <div class="form-contain">
+                    <form class="loginForm" id="loginForm">
+                      <div class="form-row">
+                        <div class="form-col-25">
+                          <label for="userName">User Name:</label>
+                        </div>
+                        <div class="form-col-75">
+                          <input
+                            name="userName"
+                            id="userName"
+                            type="text"
+                            placeholder="Enter your user name"
+                            required
+                          />
+                        </div>
                       </div>
-                      <div class="form-col-75">
-                        <input name="userName" id="userName" type="text" placeholder="Enter your user name" required>
+                      <br />
+                      <div class="form-row">
+                        <div class="form-col-25"></div>
+                        <div class="form-col-75">
+                          <span class="instructions">
+                            Passwords must be at least 8 characters and contain
+                            at least 1 number, 1 capital, and 1 special
+                            character</span
+                          >
+                        </div>
                       </div>
-                    </div><br>
-                    <div class="form-row">
-                      <div class="form-col-25">
+                      <div class="form-row">
+                        <div class="form-col-25">
+                          <label for="userPassword">Password:</label>
+                        </div>
+                        <div class="form-col-75">
+                          <input
+                            name="userPassword"
+                            id="userPassword"
+                            type="password"
+                            placeholder="Enter your password"
+                            pattern="(?=^.{8,}$)(?=.*\d)(?=.*\W+)(?![.\n])(?=.*[A-Z])(?=.*[a-z]).*$"
+                            required
+                          />
+                        </div>
                       </div>
-                      <div class="form-col-75">
-                        <span class="instructions">
-                          Passwords must be at least 8 characters
-                          and contain at least 1 number,
-                          1 capital, and 1 special character</span>
+                      <div class="form-row">
+                        <button
+                          type="submit"
+                          class="button"
+                          name="submit"
+                          id="logbtn"
+                          value="Login"
+                        >
+                          Login
+                        </button>
+                        <!-- Add the action name & value pair -->
+                        <!-- <input type="hidden" name="action" value="login-user"> -->
                       </div>
-                    </div>
-                    <div class="form-row">
-                      <div class="form-col-25">
-                        <label for="userPassword">Password:</label>
-                      </div>
-                      <div class="form-col-75">
-                        <input name="userPassword" id="userPassword" type="password" placeholder="Enter your password"
-                          pattern="(?=^.{8,}$)(?=.*\d)(?=.*\W+)(?![.\n])(?=.*[A-Z])(?=.*[a-z]).*$" required>
-                      </div>
-                    </div>
-                    <div class="form-row">
-                      <button type="submit" class="button" name="submit" id="logbtn" value="Login">Login</button>
-                      <!-- Add the action name & value pair -->
-                      <!-- <input type="hidden" name="action" value="login-user"> -->
-                    </div>
-                  </form>
+                    </form>
+                  </div>
+                  <br />
+                  <br /><br />
+                  <p class="reg-now">
+                    Don't have a trainer account yet?
+                    <br /><br />
+                    <button class="button" id="register">
+                      <a href="../register/index.html">Register Now!</a>
+                    </button>
+                  </p>
                 </div>
-                <br>
-                <br><br>
-                <p class="reg-now">
-                  Don't have a trainer account yet?
-                  <br><br>
-                  <button class="button" id="register">
-                    <a href="/src/index.html">Register Now!</a>
-                  </button>
-                </p>
               </div>
             </div>
           </div>
         </div>
       </div>
     </div>
-  </div>
-  <footer class="footer" id="main-footer"></footer>
-  <script src="/js/login.js" type="module"></script>
-</body>
-
+    <footer class="footer" id="main-footer"></footer>
+    <script src="/js/login.js" type="module"></script>
+  </body>
 </html>
