+++ src/pokemonByType/index.html
@@ -4,7 +4,8 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Yanez | Home</title>
+    <title>Yanez | Type</title>
+    <link rel="icon" type="image/png" href="/public/images/pokemonIcon.png" />
     <link rel="stylesheet" href="/css/styles.css" />
   </head>
 
@@ -36,6 +37,6 @@
       <footer class="footer" id="main-footer"></footer>
     </div>
     <script src="../js/main.js" type="module"></script>
-    <script src="../js/externalServices.mjs" type="module"></script>
+    <script src="../js/displayPokemon.mjs" type="module"></script>
   </body>
 </html>
