+++ src/pokemonPage/index.html
@@ -5,6 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Yanez | Pokemon</title>
+    <link rel="icon" type="image/png" href="/public/images/pokemonIcon.png" />
     <link rel="stylesheet" href="/css/styles.css" />
   </head>
