+++ src/public/partials/footer.html
@@ -1,1 +1,1 @@
-    &copy; 2023 Beth Yanez - WDD330: Personal Project
+&copy; 2023 Beth Yanez - WDD330: Personal Project
